fix stuff

+++ app/src/main/java/com/example/medsafecycle/database/RemoteKeys.kt
@@ -5,7 +5,7 @@ import androidx.room.PrimaryKey
 
 @Entity(tableName = "remote_keys")
 data class RemoteKeys(
-    @PrimaryKey val id: Int,
+    @PrimaryKey val id: Long,
     val prevKey: Int?,
     val nextKey: Int?
 )
