+++ app/src/main/java/com/example/medsafecycle/database/RemoteKeysDao.kt
@@ -10,7 +10,7 @@ interface RemoteKeysDao {
     @Insert(onConflict = OnConflictStrategy.REPLACE)
     suspend fun insertAll(remoteKey: List<RemoteKeys>)
     @Query("SELECT * FROM remote_keys WHERE id = :id")
-    suspend fun getRemoteKeysId(id: Int): RemoteKeys?
+    suspend fun getRemoteKeysId(id: Long): RemoteKeys?
     @Query("DELETE FROM remote_keys")
     suspend fun deleteRemoteKeys()
 }
